#99: Rename Transaction namespace to Transactions
