+++ Api/ShoppingList.Api.Infrastructure/Common/Transactions/Transaction.cs
@@ -4,7 +4,7 @@ using System.Data.Common;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace ProjectHermes.ShoppingList.Api.Infrastructure.Common.Transaction
+namespace ProjectHermes.ShoppingList.Api.Infrastructure.Common.Transactions
 {
     public class Transaction : ITransaction
     {
